refactor: 루트 프로젝트명 변경
