+++ muckpot-api/src/test/kotlin/com/yapp/muckpot/ApplicationTests.kt
@@ -1,4 +1,4 @@
-package com.yapp.web1be
+package com.yapp.muckpot
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
