+++ muckpot-domain/src/main/kotlin/com/yapp/muckpot/test/TestRepository.kt
@@ -1,4 +1,4 @@
-package com.yapp.web1be.test
+package com.yapp.muckpot.test
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
